optimize memory access (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@212 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ bios/bios02.c
@@ -17,7 +17,7 @@ static const char msg4[] = " MEMORY";
 static UINT16 textout(UINT16 ptr, const char *str) {
 
 	while(*str) {
-		i286_membyte_write(0xa000, ptr, *str);
+		i286_membyte_write(0xa000, ptr, (BYTE)*str);
 		i286_membyte_write(0xa200, ptr, 0x43);
 		str++;
 		ptr += 2;
